requirements.txt

flask>=2.0
flask-cors>=3.0
tensorflow>=2.9
numpy
opencv-python-headless
pillow
